pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "open-cadre"
version = "0.1.0"
description = "Open-source AI team platform for data engineering"
requires-python = ">=3.10"
license = "Apache-2.0"
readme = "README.md"
authors = [
    { name = "OpenCadre Contributors" },
]
keywords = ["ai", "agents", "data-engineering", "dbt", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "litellm>=1.55",
    "click>=8.0",
    "rich>=13.0",
    "jinja2>=3.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
cadre = "cadre.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cadre"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
